refactor: add operate importer 8_7 module
