+++ operate/importer-8_7/src/test/java/io/camunda/operate/zeebeimport/processors/fni/FNITransformerCacheTreePathTest.java
@@ -5,9 +5,9 @@
  * Licensed under the Camunda License 1.0. You may not use this file
  * except in compliance with the Camunda License 1.0.
  */
-package io.camunda.operate.zeebeimport.v8_6.processors.fni;
+package io.camunda.operate.zeebeimport.processors.fni;
 
-import static io.camunda.operate.zeebeimport.v8_6.processors.fni.FNITransformerTest.createZeebeRecord;
+import static io.camunda.operate.zeebeimport.processors.fni.FNITransformerTest.createZeebeRecord;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.ArgumentMatchers.eq;
 import static org.mockito.Mockito.times;
