+++ operate/importer/pom.xml
@@ -33,13 +33,13 @@
 
     <dependency>
       <groupId>io.camunda</groupId>
-      <artifactId>operate-importer-8_5</artifactId>
+      <artifactId>operate-importer-8_6</artifactId>
       <scope>runtime</scope>
     </dependency>
 
     <dependency>
       <groupId>io.camunda</groupId>
-      <artifactId>operate-importer-8_6</artifactId>
+      <artifactId>operate-importer-8_7</artifactId>
       <scope>runtime</scope>
     </dependency>
 
@@ -153,7 +153,7 @@
         <artifactId>maven-dependency-plugin</artifactId>
         <configuration>
           <usedDependencies>
-            <dependency>io.camunda:operate-importer-8_5</dependency>
+            <dependency>io.camunda:operate-importer-8_6</dependency>
           </usedDependencies>
         </configuration>
       </plugin>
