+++ parent/pom.xml
@@ -595,13 +595,13 @@
 
       <dependency>
         <groupId>io.camunda</groupId>
-        <artifactId>operate-importer-8_5</artifactId>
+        <artifactId>operate-importer-8_6</artifactId>
         <version>${project.version}</version>
       </dependency>
 
       <dependency>
         <groupId>io.camunda</groupId>
-        <artifactId>operate-importer-8_6</artifactId>
+        <artifactId>operate-importer-8_7</artifactId>
         <version>${project.version}</version>
       </dependency>
